+++ NEWS.md
@@ -45,6 +45,11 @@
  * Add support for domain-specific CRAN package **partykit**, e.g. `cf
    <- cforest(dist ~ speed, data = cars) |> futurize()`.
 
+ * Add support for domain-specific CRAN package **riskRegression**,
+   e.g. `sc <- Score(list("CSC" = fit), data = d, formula =
+   Hist(time, event) ~ 1, times = 5, B = 100, split.method =
+   "bootcv") |> futurize()`.
+
  * Add support for domain-specific Bioconductor package **scater**,
    e.g. `sce <- runPCA(sce) |> futurize()`.